+++ PurchaseSaleRecord.h
@@ -21,35 +21,32 @@ public:
     void load(std::ifstream &fin);
 
     // Computes the profit/loss for the PurchaseSaleRecord
-    double computeProfitLoss();
+    int computeProfitLoss();
 
     std::string getCompany() const { return _company; }
 
-    double getPurchaseDate() const { return _purchaseDate; }
+    int getPurchaseDate() const { return _purchaseDate; }
 
-    double getPurchasePrice() const { return _purchasePrice; }
+    int getPurchasePrice() const { return _purchasePrice; }
 
-    double getPurchaseFee() const { return _purchaseFee; }
+    int getPurchaseFee() const { return _purchaseFee; }
 
-    double getSaleDate() const { return _saleDate; }
+    int getSaleDate() const { return _saleDate; }
 
-    double getSalePrice() const { return _salePrice; }
+    int getSalePrice() const { return _salePrice; }
 
-    double getSaleFee() const { return _saleFee; }
+    int getSaleFee() const { return _saleFee; }
 
 private:
     std::string _company;
 
-    /* The following data members are stored as doubles rather than ints due to the fact that
-     * the integers involved have the potential to be very large. Better safe than sorry.*/
-    // TODO: update UML diagram with new double values
-    double _numShares;
-    double _purchaseDate;
-    double _purchasePrice;
-    double _purchaseFee;
-    double _saleDate;
-    double _salePrice;
-    double _saleFee;
+    int _numShares;
+    int _purchaseDate;
+    int _purchasePrice;
+    int _purchaseFee;
+    int _saleDate;
+    int _salePrice;
+    int _saleFee;
 };
